Put geometry in a separate file for easier static assert testing
